work_detection/requirements.txt
absl-py==2.1.0
filelock==3.16.1
fsspec==2024.10.0
grpcio==1.67.1
Jinja2==3.1.4
joblib==1.4.2
Markdown==3.7
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==24.1
pillow==11.0.0
protobuf==5.28.3
python-dotenv==1.0.1
PyYAML==6.0.2
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
torch==2.5.0
torchvision==0.20.0
tqdm==4.66.6
triton==3.1.0
typing_extensions==4.12.2
Werkzeug==3.0.6
